+++ packages/core/src/StateMachine.ts
@@ -65,6 +65,7 @@ export class StateMachine<
   TTag extends string,
   TInput,
   TOutput,
+  TEmitted extends EventObject = EventObject, // TODO: remove default
   TResolvedTypesMeta = ResolveTypegenMeta<
     TypegenDisabled,
     NoInfer<TEvent>,
@@ -72,14 +73,16 @@ export class StateMachine<
     TAction,
     TGuard,
     TDelay,
-    TTag
+    TTag,
+    TEmitted
   >
 > implements
     ActorLogic<
       MachineSnapshot<TContext, TEvent, TChildren, TStateValue, TTag, TOutput>,
       TEvent,
       TInput,
-      AnyActorSystem
+      AnyActorSystem,
+      TEmitted
     >
 {
   /**
@@ -152,9 +155,9 @@ export class StateMachine<
 
     if (
       isDevelopment &&
-      !this.root.output &&
+      !('output' in this.root) &&
       Object.values(this.states).some(
-        (state) => state.type === 'final' && !!state.output
+        (state) => state.type === 'final' && 'output' in state
       )
     ) {
       console.warn(
@@ -190,6 +193,7 @@ export class StateMachine<
     TTag,
     TInput,
     TOutput,
+    TEmitted,
     TResolvedTypesMeta
   > {
     const { actions, guards, actors, delays } = this.implementations;
@@ -259,7 +263,7 @@ export class StateMachine<
       TOutput
     >,
     event: TEvent,
-    actorScope: ActorScope<typeof snapshot, TEvent>
+    actorScope: ActorScope<typeof snapshot, TEvent, AnyActorSystem, TEmitted>
   ): MachineSnapshot<TContext, TEvent, TChildren, TStateValue, TTag, TOutput> {
     return macrostep(snapshot, event, actorScope).snapshot as typeof snapshot;
   }
@@ -345,7 +349,9 @@ export class StateMachine<
   public getInitialSnapshot(
     actorScope: ActorScope<
       MachineSnapshot<TContext, TEvent, TChildren, TStateValue, TTag, TOutput>,
-      TEvent
+      TEvent,
+      AnyActorSystem,
+      TEmitted
     >,
     input?: TInput
   ): MachineSnapshot<TContext, TEvent, TChildren, TStateValue, TTag, TOutput> {
@@ -445,7 +451,9 @@ export class StateMachine<
     snapshot: Snapshot<unknown>,
     _actorScope: ActorScope<
       MachineSnapshot<TContext, TEvent, TChildren, TStateValue, TTag, TOutput>,
-      TEvent
+      TEvent,
+      AnyActorSystem,
+      TEmitted
     >
   ): MachineSnapshot<TContext, TEvent, TChildren, TStateValue, TTag, TOutput> {
     const children: Record<string, AnyActorRef> = {};