+++ packages/core/src/createActor.ts
@@ -14,8 +14,10 @@ import { AnyActorSystem, Clock, createSystem } from './system.ts';
 import type {
   ActorScope,
   AnyActorLogic,
+  AnyActorRef,
   ConditionalRequired,
   DoneActorEvent,
+  EmittedFrom,
   EventFromLogic,
   InputFrom,
   IsNotNever,
@@ -69,7 +71,8 @@ const defaultOptions = {
  * An Actor is a running process that can receive events, send events and change its behavior based on the events it receives, which can cause effects outside of the actor. When you run a state machine, it becomes an actor.
  */
 export class Actor<TLogic extends AnyActorLogic>
-  implements ActorRef<SnapshotFrom<TLogic>, EventFromLogic<TLogic>>
+  implements
+    ActorRef<SnapshotFrom<TLogic>, EventFromLogic<TLogic>, EmittedFrom<TLogic>>
 {
   /**
    * The current internal state of the actor.
@@ -91,21 +94,30 @@ export class Actor<TLogic extends AnyActorLogic>
   );
 
   private observers: Set<Observer<SnapshotFrom<TLogic>>> = new Set();
+  private eventListeners: Map<
+    string,
+    Set<(emittedEvent: EmittedFrom<TLogic>) => void>
+  > = new Map();
   private logger: (...args: any[]) => void;
 
   /** @internal */
   public _processingStatus: ProcessingStatus = ProcessingStatus.NotStarted;
 
   // Actor Ref
-  public _parent?: ActorRef<any, any>;
+  public _parent?: AnyActorRef;
   /** @internal */
   public _syncSnapshot?: boolean;
-  public ref: ActorRef<SnapshotFrom<TLogic>, EventFromLogic<TLogic>>;
+  public ref: ActorRef<
+    SnapshotFrom<TLogic>,
+    EventFromLogic<TLogic>,
+    EmittedFrom<TLogic>
+  >;
   // TODO: add typings for system
   private _actorScope: ActorScope<
     SnapshotFrom<TLogic>,
     EventFromLogic<TLogic>,
-    any
+    AnyActorSystem,
+    EmittedFrom<TLogic>
   >;
 
   private _systemId: string | undefined;
@@ -178,6 +190,15 @@ export class Actor<TLogic extends AnyActorLogic>
           );
         }
         (child as any)._stop();
+      },
+      emit: (emittedEvent) => {
+        const listeners = this.eventListeners.get(emittedEvent.type);
+        if (!listeners) {
+          return;
+        }
+        for (const handler of Array.from(listeners)) {
+          handler(emittedEvent);
+        }
       }
     };
 
@@ -393,6 +414,25 @@ export class Actor<TLogic extends AnyActorLogic>
     return {
       unsubscribe: () => {
         this.observers.delete(observer);
+      }
+    };
+  }
+
+  public on<TType extends EmittedFrom<TLogic>['type']>(
+    type: TType,
+    handler: (emitted: EmittedFrom<TLogic> & { type: TType }) => void
+  ): Subscription {
+    let listeners = this.eventListeners.get(type);
+    if (!listeners) {
+      listeners = new Set();
+      this.eventListeners.set(type, listeners);
+    }
+    const wrappedHandler = handler.bind(undefined);
+    listeners.add(wrappedHandler);
+
+    return {
+      unsubscribe: () => {
+        listeners!.delete(wrappedHandler);
       }
     };
   }