+++ packages/core/src/createMachine.ts
@@ -6,6 +6,7 @@ import {
 } from './typegenTypes.ts';
 import {
   AnyActorRef,
+  EventObject,
   AnyEventObject,
   Cast,
   InternalMachineImplementations,
@@ -115,6 +116,7 @@ export function createMachine<
   TTag extends string,
   TInput,
   TOutput extends NonReducibleUnknown,
+  TEmitted extends EventObject,
   // it's important to have at least one default type parameter here
   // it allows us to benefit from contextual type instantiation as it makes us to pass the hasInferenceCandidatesOrDefault check in the compiler
   // we should be able to remove this when we start inferring TConfig, with it we'll always have an inference candidate
@@ -131,6 +133,7 @@ export function createMachine<
       TTag,
       TInput,
       TOutput,
+      TEmitted,
       TTypesMeta
     >;
     schemas?: unknown;
@@ -144,6 +147,7 @@ export function createMachine<
     TTag,
     TInput,
     TOutput,
+    TEmitted,
     TTypesMeta
   >,
   implementations?: InternalMachineImplementations<
@@ -155,7 +159,8 @@ export function createMachine<
       TAction,
       TGuard,
       TDelay,
-      TTag
+      TTag,
+      TEmitted
     >
   >
 ): StateMachine<
@@ -177,14 +182,25 @@ export function createMachine<
       TAction,
       TGuard,
       TDelay,
-      TTag
+      TTag,
+      TEmitted
     >['resolved'],
     'tags'
   > &
     string,
   TInput,
   TOutput,
-  ResolveTypegenMeta<TTypesMeta, TEvent, TActor, TAction, TGuard, TDelay, TTag>
+  TEmitted,
+  ResolveTypegenMeta<
+    TTypesMeta,
+    TEvent,
+    TActor,
+    TAction,
+    TGuard,
+    TDelay,
+    TTag,
+    TEmitted
+  >
 > {
   return new StateMachine<
     any,
@@ -198,6 +214,7 @@ export function createMachine<
     any,
     any,
     any,
+    any, // TEmitted
     any
   >(config as any, implementations as any);
 }