+++ packages/core/src/setup.ts
@@ -9,6 +9,7 @@ import {
   Cast,
   ConditionalRequired,
   DelayConfig,
+  EventObject,
   Invert,
   IsNever,
   MachineConfig,
@@ -27,44 +28,43 @@ type ToParameterizedObject<
     string,
     ParameterizedObject['params'] | undefined
   >
-> = Values<{
-  [K in keyof TParameterizedMap & string]: {
-    type: K;
-    params: TParameterizedMap[K];
-  };
-}>;
-
-type DefaultToUnknownActorLogic<
-  TActors extends Record<string, UnknownActorLogic>
-> =
-  // if `keyof TActors` is `never` then it means that both `children` and `actors` were not supplied
-  // `never` comes from the default type of the `TChildrenMap` type parameter
-  // in such a case we "replace" `TActors` with a more traditional~ constraint
-  // one that doesn't depend on `Values<TChildrenMap>`
-  IsNever<keyof TActors> extends true
-    ? Record<string, UnknownActorLogic>
-    : TActors;
+> = // `silentNeverType` to `never` conversion (explained in `ToProvidedActor`)
+  IsNever<TParameterizedMap> extends true
+    ? never
+    : Values<{
+        [K in keyof TParameterizedMap & string]: {
+          type: K;
+          params: TParameterizedMap[K];
+        };
+      }>;
 
 // at the moment we allow extra actors - ones that are not specified by `children`
 // this could be reconsidered in the future
 type ToProvidedActor<
   TChildrenMap extends Record<string, string>,
-  TActors extends Record<Values<TChildrenMap>, UnknownActorLogic>,
-  TResolvedActors extends Record<
-    string,
-    UnknownActorLogic
-  > = DefaultToUnknownActorLogic<TActors>
-> = Values<{
-  [K in keyof TResolvedActors & string]: {
-    src: K;
-    logic: TResolvedActors[K];
-    id: IsNever<TChildrenMap> extends true
-      ? string | undefined
-      : K extends keyof Invert<TChildrenMap>
-        ? Invert<TChildrenMap>[K] & string
-        : string | undefined;
-  };
-}>;
+  TActors extends Record<string, UnknownActorLogic>
+> =
+  // this essentially is meant to convert a leaked `silentNeverType` to the true `never` type
+  // it shouldn't be observable but here we are
+  // we don't want to lock inner inferences for our actions with types containing this type
+  // it's used in inner inference contexts when the outer one context doesn't have inference candidates for a type parameter
+  // because it leaks here, without this condition it manages to create an inferrable type that contains it
+  // the `silentNeverType` is non-inferrable itself and that usually means that a containing object is non-inferrable too
+  // that doesn't happen here though. However, we actually want to infer a true `never` here so our actions can't use unknown actors
+  // for that reason it's important to do the conversion here because we want to map it to something that is actually inferrable
+  IsNever<TActors> extends true
+    ? never
+    : Values<{
+        [K in keyof TActors & string]: {
+          src: K;
+          logic: TActors[K];
+          id: IsNever<TChildrenMap> extends true
+            ? string | undefined
+            : K extends keyof Invert<TChildrenMap>
+              ? Invert<TChildrenMap>[K] & string
+              : string | undefined;
+        };
+      }>;
 
 type _GroupStateKeys<
   T extends StateSchema,
@@ -103,17 +103,28 @@ type ToStateValue<T extends StateSchema> = T extends {
             : never)
   : {};
 
+type RequiredSetupKeys<TChildrenMap> = IsNever<keyof TChildrenMap> extends true
+  ? never
+  : 'actors';
+
 export function setup<
   TContext extends MachineContext,
   TEvent extends AnyEventObject, // TODO: consider using a stricter `EventObject` here
-  TActors extends Record<Values<TChildrenMap>, UnknownActorLogic>,
-  TActions extends Record<string, ParameterizedObject['params'] | undefined>,
-  TGuards extends Record<string, ParameterizedObject['params'] | undefined>,
-  TDelay extends string,
-  TTag extends string,
-  TInput,
-  TOutput extends NonReducibleUnknown,
-  TChildrenMap extends Record<string, string> = never
+  TActors extends Record<string, UnknownActorLogic> = {},
+  TChildrenMap extends Record<string, string> = {},
+  TActions extends Record<
+    string,
+    ParameterizedObject['params'] | undefined
+  > = {},
+  TGuards extends Record<
+    string,
+    ParameterizedObject['params'] | undefined
+  > = {},
+  TDelay extends string = never,
+  TTag extends string = string,
+  TInput = NonReducibleUnknown,
+  TOutput extends NonReducibleUnknown = NonReducibleUnknown,
+  TEmitted extends EventObject = EventObject
 >({
   schemas,
   actors,
@@ -122,9 +133,21 @@ export function setup<
   delays
 }: {
   schemas?: unknown;
-  types?: SetupTypes<TContext, TEvent, TChildrenMap, TTag, TInput, TOutput>;
+  types?: SetupTypes<
+    TContext,
+    TEvent,
+    TChildrenMap,
+    TTag,
+    TInput,
+    TOutput,
+    TEmitted
+  >;
   actors?: {
-    [K in keyof TActors]: TActors[K];
+    // union here enforces that all configured children have to be provided in actors
+    // it makes those values required here
+    [K in keyof TActors | Values<TChildrenMap>]: K extends keyof TActors
+      ? TActors[K]
+      : never;
   };
   actions?: {
     [K in keyof TActions]: ActionFunction<
@@ -135,7 +158,8 @@ export function setup<
       ToProvidedActor<TChildrenMap, TActors>,
       ToParameterizedObject<TActions>,
       ToParameterizedObject<TGuards>,
-      TDelay
+      TDelay,
+      TEmitted
     >;
   };
   guards?: {
@@ -154,6 +178,8 @@ export function setup<
       TEvent
     >;
   };
+} & {
+  [K in RequiredSetupKeys<TChildrenMap>]: unknown;
 }): {
   createMachine: <
     const TConfig extends MachineConfig<
@@ -166,6 +192,7 @@ export function setup<
       TTag,
       TInput,
       TOutput,
+      TEmitted,
       ResolveTypegenMeta<
         TypegenDisabled,
         TEvent,
@@ -173,7 +200,8 @@ export function setup<
         ToParameterizedObject<TActions>,
         ToParameterizedObject<TGuards>,
         TDelay,
-        TTag
+        TTag,
+        TEmitted
       >
     >
   >(
@@ -193,6 +221,7 @@ export function setup<
     TTag,
     TInput,
     TOutput,
+    TEmitted,
     ResolveTypegenMeta<
       TypegenDisabled,
       TEvent,
@@ -200,7 +229,8 @@ export function setup<
       ToParameterizedObject<TActions>,
       ToParameterizedObject<TGuards>,
       TDelay,
-      TTag
+      TTag,
+      TEmitted
     >
   >;
 } {