+++ packages/xstate-test/test/dieHard.test.ts
@@ -9,29 +9,6 @@ describe('die hard example', () => {
     five: number;
   }
 
-  const pour3to5 = assign<DieHardContext>(({ context }) => {
-    const poured = Math.min(5 - context.five, context.three);
-
-    return {
-      three: context.three - poured,
-      five: context.five + poured
-    };
-  });
-  const pour5to3 = assign<DieHardContext>(({ context }) => {
-    const poured = Math.min(3 - context.three, context.five);
-
-    const res = {
-      three: context.three + poured,
-      five: context.five - poured
-    };
-
-    return res;
-  });
-  const fill3 = assign<DieHardContext>({ three: 3 });
-  const fill5 = assign<DieHardContext>({ five: 5 });
-  const empty3 = assign<DieHardContext>({ three: 0 });
-  const empty5 = assign<DieHardContext>({ five: 0 });
-
   class Jugs {
     public version = 0;
     public three = 0;
@@ -79,22 +56,38 @@ describe('die hard example', () => {
             },
             on: {
               POUR_3_TO_5: {
-                actions: pour3to5
+                actions: assign(({ context }) => {
+                  const poured = Math.min(5 - context.five, context.three);
+
+                  return {
+                    three: context.three - poured,
+                    five: context.five + poured
+                  };
+                })
               },
               POUR_5_TO_3: {
-                actions: pour5to3
+                actions: assign(({ context }) => {
+                  const poured = Math.min(3 - context.three, context.five);
+
+                  const res = {
+                    three: context.three + poured,
+                    five: context.five - poured
+                  };
+
+                  return res;
+                })
               },
               FILL_3: {
-                actions: fill3
+                actions: assign({ three: 3 })
               },
               FILL_5: {
-                actions: fill5
+                actions: assign({ five: 5 })
               },
               EMPTY_3: {
-                actions: empty3
+                actions: assign({ three: 0 })
               },
               EMPTY_5: {
-                actions: empty5
+                actions: assign({ five: 0 })
               }
             }
           },